+++ frontend/src/app/pages/exercises/exercise/shared/exercise-map/shared/simulated-region-popup/simulated-region-popup.component.ts
@@ -8,6 +8,7 @@ import type { AppState } from 'src/app/state/app.state';
 import { createSelectSimulatedRegion } from 'src/app/state/application/selectors/exercise.selectors';
 import { selectCurrentMainRole } from 'src/app/state/application/selectors/shared.selectors';
 import { openSimulationTrainerModal } from '../../../simulation/trainer-modal/open-simulation-trainer-modal';
+import { openPreviewModal } from '../../../simulation/trainer-modal/open-preview-modal';
 import { PopupService } from '../../utility/popup.service';
 
 @Component({
@@ -36,6 +37,10 @@ export class SimulatedRegionPopupComponent implements OnInit {
     openInModal() {
         this.popupService.closePopup();
         openSimulationTrainerModal(this.modalService, this.simulatedRegionId);
+    }
+
+    openPreviewModal() {
+        openPreviewModal(this.modalService, this.simulatedRegionId);
     }
 
     public closePopup() {