+++ frontend/src/app/pages/exercises/exercise/shared/simulation/trainer-modal/preview-card/simulated-region-preview-card.component.ts
@@ -1,0 +1,16 @@
+import { Component, input, output } from '@angular/core';
+
+@Component({
+    selector: 'app-simulated-region-preview-card',
+    templateUrl: './simulated-region-preview-card.component.html',
+    styleUrls: ['./simulated-region-preview-card.component.scss'],
+    standalone: false,
+})
+export class SimulatedRegionPreviewCardComponent {
+    readonly elementMousedown = output<MouseEvent>();
+
+    dataCy = input<string>('');
+    title = input.required<string>();
+    imageUrl = input.required<string>();
+    darkBackground = input(false);
+}